requirements.txt
flask>=2.0.0
ccxt>=3.0.0
pandas>=1.3.0
matplotlib>=3.4.0
sqlalchemy>=1.4.0
requests>=2.26.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
pytest>=6.2.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
sentencepiece>=0.1.99
apscheduler>=3.10.0
numpy>=1.21.0
slack-sdk>=3.19.0
ta>=0.10.2
backtrader>=1.9.76.123
plotly>=5.13.0
scipy>=1.7.0